Fix contracts build.
Contracts were not compiling after merging:
https://github.com/slockit/DAO/pull/146

This fixes the build by adding the extra Token EIP variables from the
places they were missing and also using `uint8` for the decimal places everywhere.

+++ TokenCreation.sol
@@ -88,7 +88,7 @@ contract TokenCreation is TokenCreationInterface, Token {
         address _privateCreation,
         string _tokenName,
         string _tokenSymbol,
-        uint _decimalPlaces) {
+        uint8 _decimalPlaces) {
 
         closingTime = _closingTime;
         minTokensToCreate = _minTokensToCreate;
